test: RoomSetting 테스트와 RoomTest 분리 #215
